Minor change to logging.
Consolidated 4 logging events that happen at the same time to a single
event, and changed the level from 'error' to 'warning'. An exception is
also raised at this point, and if an error is needed instead of a warning,
it should be issued by the code that handles the exception.